requirements.txt
supabase-py
Flask
Flask-Cors
yt-dlp